Extract city object preparation
